Refactor HTTP get requesting code
